Split converting pdfs to text and text mining into separate scripts
These are naturaly separate chunks of code.

This is the first step in making it easier to automatically rerun everything
quickly after adding new data. Following this split we can shift the convert
script to running on a pdf by pdf basis and then automate running it on all
pdfs with make so that we don't rerun on pdfs that have already been converted
and that remain unchanged.
